Миллион изменений

+++ IGameObject/IGameObject.cpp
@@ -2,7 +2,7 @@
 
 sf::Texture IGameObject::texture;
 
-IGameObject :: IGameObject(double x, double y) : dx(x), dy(y) {
+IGameObject :: IGameObject(float x, float y) : dx(x), dy(y) {
     sprite.setTexture((IGameObject::texture));
     isDestroyed = false;
 }
